fix(ats,research): fix Lever float salary, resolve 6 invalid-token records
Lever's salaryRange.min/max are sometimes floats (e.g. 22.5); change struct
fields from int64 to float64 and cast to int64 on return. Fixes parse failures
for boards like Ridwell that advertise fractional hourly rates.

Re-annotated 6 GeekWire 200 invalid-token records: Helion, Hiya, and Qumulo
moved from Greenhouse to Ashby; Textio moved from Lever to Greenhouse; Ridwell
board token is case-sensitive (Ridwell, not ridwell); Tinuiti Workday host is
wd12 not wd1. Also marked Skilljar (acquired by Gainsight), Formidable (acquired
by NearForm), and Boundless rank 190 (duplicate source entry) to terminal status.
Verified count now 73, invalid-token down to 16.

+++ internal/ats/lever.go
@@ -89,10 +89,10 @@ type leverJob struct {
 // leverSalaryRange is Lever's structured compensation block. Pointer on the
 // parent so a wire-absent field stays distinct from a zero-valued struct.
 type leverSalaryRange struct {
-	Min      int64  `json:"min"`
-	Max      int64  `json:"max"`
-	Currency string `json:"currency"`
-	Interval string `json:"interval"`
+	Min      float64 `json:"min"`
+	Max      float64 `json:"max"`
+	Currency string  `json:"currency"`
+	Interval string  `json:"interval"`
 }
 
 // leverIntervalAliases maps Lever's wire `salaryRange.interval` values to the
@@ -333,7 +333,7 @@ func normalizeLeverSalary(s *leverSalaryRange) (min int64, max int64, currency s
 		slog.Warn("[lever] invalid salary range", "min", s.Min, "max", s.Max)
 		return 0, 0, "", "", false
 	}
-	return s.Min, s.Max, currency, period, true
+	return int64(s.Min), int64(s.Max), currency, period, true
 }
 
 // isThreeUpperLetters reports whether s is exactly three ASCII uppercase
